Add debug logging for some IMAP commands

+++ fsstore_test.go
@@ -66,7 +66,7 @@ func initTestBackend() backendtests.Backend {
 
 func cleanBackend(bi backendtests.Backend) {
 	b := bi.(*Backend)
-	if os.Getenv("PRESERVE_DB") != "1" {
+	if os.Getenv("PRESERVE_DB") != "1" && os.Getenv("PRESERVE_SQLITE3_DB") != "1" {
 		// Remove things manually in the right order so we will not hit
 		// foreign key constraint when dropping tables.
 		if _, err := b.DB.Exec(`DELETE FROM msgs`); err != nil {
